v 0.1
* fix some converts float -> int and another
* add background game
* add board
* add colorsite
* new fonts
* add new dir buttons
* add some structure game classes

+++ monopoly2d-sfml/monopoly2d-sfml/GameWindow.cpp
@@ -38,7 +38,7 @@ void GameWindow::close()
 
 void GameWindow::draw()
 {
-	for (int i = 0; i < drawable_objects.size(); i++)
+	for (size_t i = 0; i < drawable_objects.size(); i++)
 	{
 		window->draw(*drawable_objects.at(i));
 	}
